Move MockTradeWallet to new 'protocol::mocks' module
Place the mock buyer/seller trade wallets ('ProtocolWalletAPI' impls)
into their own module, leaving 'protocol::psbt' as a private utility
module, used exclusively by 'protocol::transaction' for building/merging
the half-deposit PSBTs and validating/extracting the signed Deposit &
Custom Payout Txs from their respective PSBTs. (The 'psbt::Redact' trait
is also being used by 'MockTradeWallet::sign_selected_inputs', however.)

The new 'mocks' module is currently part of the production crate, for
use of the mock trade wallets by 'rpc::protocol' in the 'Musig' gRPC
service implementation, but will probably eventually be feature-gated or
made "#![cfg(test)]".
